Simple communication between React and Node.
