+++ src/index.js
@@ -1,16 +1,1 @@
-import React, { Component } from 'react';
-import ReactDOM from 'react-dom';
-
-class App extends Component {
-  render(){
-    return(<div>
-      <h1>HyPhy</h1>
-      <p>Hypothesis testing using phylogenies</p>
-    </div>);
-  }
-}
-
-ReactDOM.render(
-  <App />,
-  document.body.appendChild(document.createElement('div'))
-)
+window.render_app = require('./app');
